Update ByteBufferReader
* CoreLib(NET5.0-SDK)的部分API没有开放，SpanHelper还是只能使用.NET Core 3.1版本。

* Common/Buffer：NET5.0框架下，移除对SpanHelper的依赖。

+++ src/DotNetty.Common/Internal/Utf8Utility.Validation.cs
@@ -42,7 +42,7 @@ namespace DotNetty.Common.Internal
                 // If so, short-circuit the remainder of the method.
 
                 inputLength -= (int)numAsciiBytesCounted;
-                if (0u >= inputLength)
+                if (0u >= (uint)inputLength)
                 {
                     utf16CodeUnitCountAdjustment = 0;
                     scalarCountAdjustment = 0;
